+++ apps/pie-docs/src/componentStatusData.js
@@ -1,4 +1,4 @@
-const webComponentStatuses = require('../component-statuses.json');
+const webComponentStatuses = require('@justeattakeaway/pie-monorepo-utils/dist/component-statuses.json');
 const resourceTypes = require('./_data/resourceTypes');
 const statusTypes = require('./_data/statusTypes');
 
@@ -1435,16 +1435,56 @@ const rows = [
         },
         {
             resource: resourceTypes.WEB_COMPONENTS,
-            status: statusTypes.PLANNED,
+            link: 'https://webc.pie.design/?path=/story/radio--default',
+            status: webComponentStatuses['pie-radio'],
         },
         {
             resource: resourceTypes.VUE,
-            link: 'https://vue.pie.design/?path=/story/components-molecules--alert-component',
+            link: 'https://vue.pie.design/?path=/story/components-atoms-f-form-field--radio-component',
             status: statusTypes.STABLE,
         },
         {
             resource: resourceTypes.REACT,
             link: 'https://snacks.takeaway.com/portal/components/radio/',
+            status: statusTypes.STABLE,
+        },
+        {
+            resource: resourceTypes.IOS_JUST_UI,
+            status: statusTypes.STABLE,
+        },
+        {
+            resource: resourceTypes.IOS_PIE,
+            status: statusTypes.STABLE,
+        },
+        {
+            resource: resourceTypes.ANDROID_PIE,
+            status: statusTypes.STABLE,
+        }
+    ],
+    [
+        { componentName: 'Radio Group' },
+        {
+            resource: resourceTypes.COMPONENT,
+            link: 'https://www.figma.com/design/pPSC73rPin4csb8DiK1CRr/%E2%9C%A8-%5BCore%5D-Web-Components-%5BPIE-3%5D?node-id=198-33663',
+            status: statusTypes.STABLE,
+        },
+        {
+            resource: resourceTypes.DOCUMENTATION,
+            link: 'https://www.figma.com/design/OOgnT2oNMdGFytj5AanYvt/%5BCore%5D-Web-Component-Documentation-%5BPIE-3%5D?node-id=29-22',
+            status: statusTypes.STABLE,
+        },
+        {
+            resource: resourceTypes.WEB_COMPONENTS,
+            link: 'https://webc.pie.design/?path=/story/radio-group--default',
+            status: webComponentStatuses['pie-radio-group'],
+        },
+        {
+            resource: resourceTypes.VUE,
+            status: statusTypes.NOT_APPLICABLE,
+        },
+        {
+            resource: resourceTypes.REACT,
+            link: 'https://snacks.takeaway.com/portal/components/radio-group/',
             status: statusTypes.STABLE,
         },
         {
@@ -1937,7 +1977,7 @@ const rows = [
         },
         {
             resource: resourceTypes.WEB_COMPONENTS,
-            status: statusTypes.PLANNED,
+            status: webComponentStatuses['pie-textarea'],
         },
         {
             resource: resourceTypes.VUE,
@@ -2009,11 +2049,13 @@ const rows = [
         },
         {
             resource: resourceTypes.DOCUMENTATION,
-            status: statusTypes.PLANNED,
+            link: 'https://www.figma.com/design/pPSC73rPin4csb8DiK1CRr/branch/sjkhCbCcv7t4QkDrz7g0Xx/%E2%9C%A8-%5BCore%5D-Web-Components-%5BPIE-3%5D?node-id=7934-11960&node-type=frame&t=SxRV1mF3FT7S4lAP-0',
+            status: statusTypes.STABLE,
         },
         {
             resource: resourceTypes.WEB_COMPONENTS,
-            status: statusTypes.PLANNED,
+            link: 'https://webc.pie.design/?path=/story/toast--neutral',
+            status: webComponentStatuses['pie-toast'],
         },
         {
             resource: resourceTypes.VUE,