+++ packages/components/pie-tag/src/defs.ts
@@ -1,7 +1,8 @@
 import { type ComponentDefaultProps } from '@justeattakeaway/pie-webc-core';
 
-export const variants = ['neutral-alternative', 'neutral', 'outline', 'ghost', 'blue', 'green', 'yellow', 'red', 'brand', 'brand-03', 'brand-04', 'brand-06'] as const;
+export const variants = ['neutral-alternative', 'neutral', 'outline', 'ghost', 'information', 'success', 'error', 'brand-02', 'brand-03', 'brand-04', 'brand-05', 'brand-06'] as const;
 export const sizes = ['small', 'large'] as const;
+export const iconPlacements = ['leading', 'trailing'] as const;
 
 export interface TagProps {
     /**
@@ -10,7 +11,7 @@ export interface TagProps {
     variant?: typeof variants[number];
 
     /**
-     * When true, the "green", "yellow", "red", "blue" and "neutral" variants change their styles and become bolder
+     * When true, the "information", "success", "error", "brand-05", and "neutral" variants change their styles and become bolder
      */
     isStrong?: boolean;
 
@@ -29,6 +30,11 @@ export interface TagProps {
      * What size the tag should be.
      */
     size?: typeof sizes[number];
+
+    /**
+     * The placement of the icon slot such as leading (default) or trailing. Available only if `isInteractive` is set to true.
+     */
+    iconPlacement?: typeof iconPlacements[number];
 }
 
 export type DefaultProps = ComponentDefaultProps<TagProps>;
@@ -39,4 +45,5 @@ export const defaultProps: DefaultProps = {
     isInteractive: false,
     disabled: false,
     size: 'large',
+    iconPlacement: 'leading',
 };